scheduler/requirements.txt
MozillaPulse==1.0.1
PyHawk-with-a-single-extra-commit==0.1.5
PyYAML==3.11
SQLAlchemy==0.9.8
Twisted==15.0.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
buildtools==1.0.6
docopt==0.6.2
furl==0.4.4
kombu==3.0.24
orderedmultidict==0.7.4
python-dateutil==2.4.0
pytz==2014.10
requests==2.4.3
simplejson==3.6.5
six==1.9.0
taskcluster==0.0.9
wsgiref==0.1.2
zope.interface==4.1.2
